refactor/made some changes to the file struc to follow snakecase
